squash everything into IdeHighlightStencil
remove note and feature classes
move everything related to the highlight stencil into org.alice.ide.highlight
